backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
fastapi = "^0.63.0"
SQLAlchemy = "^1.3.23"
python-jose = "^3.2.0"
passlib = "^1.7.4"
celery = "^5.0.5"
pandas = "^1.2.1"
numpy = "^1.20.0"
uvicorn = "^0.13.3"
tenacity = "^6.3.1"
emails = "^0.6"
email-validator = "^1.1.2"
openpyxl = "^3.0.6"
psutil = "^5.8.0"
alembic = "^1.5.4"
python-socketio = "^5.0.4"
psycopg2-binary = "^2.8.6"
redis = "^3.5.3"
bcrypt = "^3.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
